-- 07.08.2020 ----

+++ ShaperLayer.py
@@ -30,7 +30,7 @@ class ShaperLayer(tf.keras.layers.Layer):
 
     def __init__(self, input_dim):
         super(ShaperLayer, self).__init__()
-        self.total = tf.Variable(initial_value=tf.zeros((input_dim,)), trainable=False, dtype='float32')
+        self.total = tf.Variable(initial_value=tf.zeros((input_dim,)), trainable=False, dtype='float64')
 
     def call(self, inputs):
         return tf.convert_to_tensor(self.concat_result(inputs.numpy()))
